Properly handle XCM control interface fds
When the 'pafd' process wakes up from epoll_wait(), it will receive a
list of active fds from the kernel. The server has each fd tied to an
event loop source, which in turn has a handler callback function.

The event loop didn't properly handle the case where a client handler
would interact with its XCM connection in such a way that the XCM
socket would close a socket, and the next handler would open another
file object with the same fd number, but references a different kernel
file object.

For this to happen, the pafd process need to be under high load and
the XCM control interface (xcmctl) would need to be actively used.

In addition, this change address a related bug with the same symptoms,
where the core server code failed to update the fds registered in the
event loop. This problem also only occurs per above.

Test cases are added to reproduce these bugs, including a newly
created (and currently very small) event loop test suite.

Change-Id: I89c923217ba4834c6a58a34b68622c3d8a1c3431