Avoid running out of fds in multi-domain test case
Reduce the number of test domains from max 32 to max 16 to avoid
running out of server fds on a system configured for max 1024 files
per process (the default on most Linux distros).

This change affect the test case only. The server still supports an
arbitrary number of domains per instance.

Change-Id: Ic6e1b07eb2fa88d0cd06faf1145e477f7d8b6ea8