Removed XCM wrapper non-optimization
The Pathfinder Python XCM wrapper had logic to keep track if any
socket state-alternating functions has been called. This per-socket
flag is checked in a want() call. In case the socket is not "dirty",
the previous result is returned, saving the interpreter for doing an
xcm_want() call (via ctypes).

This optimization is a non-optimization for the Pathfinder server,
since it doesn't call xcm_want() on sockets that hasn't changed
(i.e. are "clean").