Improve security and introduce server configuration file
With this patch, the server supports access control, as in not
allowing anyone but the original publisher of a service to modify
it.

The server now also optionally imposes per-user and global resource
limits. Resources are client connections, services and subscriptions.
The aim of this mechanism is primarily to protect the server and
connected clients (and by extension potentially a larger system) from
denial of service (DoS) type attacks from one or more malious or
benevolent but buggy clients.

What constitute a user depends on how a client connects to the server.
For TLS, the X509v3 Subject Key Id of the client's certificate is used
as the user id. For TCP, the client's IP address is used. Clients
using other XCM transports are currently all considered be owned by a
single, default user instance.

The subject key id is retrieve via XCM's get_attr() facility. The
"tls.peer_subject_key_id" attribute is available in XCM API version 12
or later. Older versions of XCM may still be used, which in case the
server will identify a TLS client in the same, and less secure, way as
it does for TCP-based clients.

There are no notion of a user in the wire protocol. This feature does
introduce new protocol-level failure reason codes, but this will
likely not break existing clients. An exception was 'pafc', which
wasn't prepared to handle any errors as connection establishment time
(which this patch fixes).

As a part of this change, the domain layer (i.e. sd.py) now has a
client and a user concept.

This patch changes the default total client limit from 1000 to
unlimited.

The command-line parameters and their semantics remain the
same. Command-line configuration may be combined with a configuration
file. Command-line parameters will override relevant parts of the
server configuration derived from the configuration file.

In case a configuration file is used, the server now requires the
'yaml' Python module.

Ten more test cases are added to the server test suite.