Eliminate race in test case poll usage
The 'wait()' function used by the test suite used poll() in a racy
way. The race was between signal delivery and the test case thread
entering the poll() syscall. If the signal callback would change the
criteria from false to true, this change could happen after the
'wait()' function had checked the criteria, but before the thread had
entered the poll() syscall (and thus no EINTR would be triggered).

This would occasionally cause test cases to show false negatives.

The new 'wait()' uses a file descriptor for signal delivery, which is
race free.