Improve client disconnect performance
Reduce server processing latency for handling disconnecting clients.

The connected clients were stored in a list, and the removal of a
client included a linear search, which much hurt performance when
clients were disconnected from a busy server.

Sets are now used to store client objects. In addition, the original
single list is split in two: one set for clients which has finished
the Pathfinder protocol-level hand shake, and one for those which
haven't. This speeds up the procedure of purging idle connections.

Resolves #5.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>