Make paf.client.connect blocking by default
paf.client.connect() now follows the same pattern as all the methods
in paf.client.Client and is blocking by default. If non-blocking
behavior is preferred, the caller my supply a callback function.

This change is done both for consistency and practical reasons. The
Python client API is mostly used in a blocking manner.

Resolves #4.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>