Further improve test server address reuse avoidance
Using a larger value range for the randomly generated server addresses
only in part fixed the issue of collisions. In particular, collisions
may still happen with other TCP-based services binding to '*'
(i.e. all IPv4 interfaces).

The code fragment which attempts to bind() to a server address to
verify it's not used by any other process in the system is
reintroduced.

This is purely a change to the test suite.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>