Remove non-ASCII character from README
Python 3.6 setuptools doesn't like non-ASCII characters in the
long_description field. For Pathfinder, the value of this field is
populated by reading README.md.

This issue would cause setup.py to fail on Python 3.6 and likely
earlier versions as well. On Python 3.7 and 3.8, it works.

This patch replaces the non-ASCII character with an ASCII equivalent,
with a near-identical glyph.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>