Refactor pafd script
The bulk of the functionality of the 'pafd' script is moved from the
script file itself, to a new module 'paf.daemon'. The code is split
into two parts; one function that builds the daemon configuration by
parsning command line arguments and reading the configuration file,
and one function that starts the server instance(s) with a certain
configuration.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>