Allow user hook before server start
Allow the execution of a user-supplied Python function after
configuring, but before starting the server. The hook format is
'<module>.<function>' and it's specified with the undocumented '-r'
command-line parameter. The function is supplied the list of server
objects, as its only argument.

This not-very-elegant construct is designed to be a convenient way of
integration site-local extensions (such as hooking in a custom log
handler).

Resolves #11.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>