Enforce Pathfinder protocol integer ranges
All Pathfinder protocol v2 integers are non-negative 64-bit integers,
with a maximum value of 2^63 - 1 (i.e. INT64_MAX).

This patch enforces this maximum limit in the server. It thus produces
a protocol-level if a client sends a too-large client id, for example.

The old behaviour would cause problems in case a Pathfinder client
would, in violation of the protocol, publish a service id with a
too-large service id. The server itself had no issue handling this
larger-than-allowed id, but libpaf-based clients are more strict in
what they accept. Such a client would disconnect from the server upon
receiving a subscription notification with such a too-large service id
(since the server was violating the protocol).

Test cases are added to verified this new behavior.

This is not a change of the protocol, but merely the server being more
strict.

Resolves #14.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>