Improve server responsiveness by optimizing orphan handling
When clients with a large number of published services disconnect
without first unpublishing their services (e.g. due to a client crash
or a network outage), the server processing that follows had a very
high time complexity O(N^2).

Further, while processing these services, the server would remain
unresponsive to other clients' requests.

This patch optimizes the orphan handling, which was largely
responsible for the overhead, by tracking which services are orphans,
and keeping them in temporal order.

Prior to this change, a client with 10k services (with non-zero TTL)
leaving the system would make the server spend ~35 s processing the
orphans. After this patch, on the same system, the same processing
takes ~80 ms.

This patch extends the test suite. A new test case measures the server
responsiveness when a large client disconnects, as well as when its
services are removed (i.e, when the TTL has expired). Another test
case verifying the ordering of orphan removal notification sent by the
server is added.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>