Prevent data change on same-generation republications
The server now rejects client attempts to republish services with the
same generation number, but different properties and/or
TTL. Republishing such a service is a violation of both the Pathfinder
protocol and the conceptual model with generation number-protected
service data.

This change slightly decreases performance for republish operations.

One legacy test erroneously performed such a republication, and is
fixed.

The server now use the 'same-generation-but-different' protocol-level
reason code.

Resolves #29.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>