Implement Pathfinder protocol-level keepalive
Add support in pafd for a draft of Pathfinder protocol version 3,
which adds a protocol-level liveness tracking mechanism, allowing
clients to track servers and vice versa.

Pathfinder protocol version 2 depends on the state of the transport
connection (e.g., a UNIX domain socket or a TCP connection) to
determine if a client or a server is alive and available. On TCP, TCP
keepalive is used.

This mechanism is unable to detect a number of remote peer or network
failures, such as a TCP-level relay in-between the client and server
failing to forward data without tearing down the connection. It also
couldn't detect a situation where the remote peer process was "dead in
the water" (i.e., failing to respond to a request or otherwise doing
anything useful, while also not terminating its process, keeping the
transport connection alive).

An additional issue with transport-level liveness checking is the
dependency of TCP keep alive, which does not travel through "middle
boxes" relaying on the level of TCP or TLS, such as certain Cloud
infrastructure load balancers. This force the operator to enable TCP
keepalive in the middle box.

In Pathfinder protocol version 3, the liveness tracking is moved into
the application protocol.

A new protocol command track is added, allowing for both the server to
ping the client and vice versa (although the former was already
possible with the "ping" command). The new command is of a new two-way
transaction type.

On v3 connections, server disables TCP keepalive on TCP-based XCM
transports, since it's no longer required and causes network traffic
and in-kernel TCP keepalive processing.

This commit also introduces adaptive liveness checking intervals. The
default maximum allowed idle time is 30 s, but in case a client
publishes a low-TTL service, the server-side maximum idle time is
reduced, under the assumption that low TTL implies that low state
propagation (e.g., "provider is down") latency is desirable. To avoid
excessive signaling, server never employs a max idle time lower than 2
seconds.

Processing protocol-level heartbeat messages is much more
computationally expensive than the transport-level equivalent.  For
TCP, the frequency is reduced, from two query-response pairs per
second (in the TCP keepalive case) to one query-response pair (plus
CPI ACKs) per 30 seconds.

pafc and paf.client is updated to reflect the changes. The paf.client
API is extended but is backward-compatible. However, paf.client-based
applications keeping connections idle for long periods of time may
need to enabled tracking and properly call Client.process() when the
fd is active to avoid the server timing out the connection.

pafd still support version 2 of the Pathfinder protocol, and v2-only
clients are unaffected by this change.

v2 and v3 clients may be connected to the same server and domain.

Resolves #26 and #17.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>