Add common timer manager
Introduce timer manager common to both the domain model and the server
class.

The new timer manager make the server perform better when non-uniform
service TTLs are used.

A new timer manager test suite is added.

This commit also include a change in terminology. The word "purge" is
no longer used to describe the act of dropping connections which have
failed to finish the Pathfinder protocol handshake in time, since this
term is used for other purposes when handling orphan services.

Resolves #41.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>