Update ABNF to disallow zero-length strings
The filter ABNF grammer in the Pathfinder protocol specification
allowed for zero-length keys, which in turn suggests zero-length
service property names would be allowed, which does not make a lot of
sense.

All known implementations already enforce this limitation.

The value in an equals filter may still be an empty string, as should
be allowed.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>