Make TLS credentials configurable
This patch adds support for an extended server configuration format,
aiming to provide per-server socket (or "endpoint") TLS configuration
options.

The 'addrs' key of the domain YAML dictionary is renamed 'sockets', to
better reflect the intent of the list it holds as a value. 'addrs' is
still accepted as an alias.

Each element of the 'sockets' list is now either an address (a YAML
string) or a dictionary. In the latter case, there must be a key
'addr' holding the address, and an optional key 'tls', allowed only
for TLS or UTLS type addresses.

'tls' must be a dictionary and may contain any of 'cert', 'key' or
'tc'. All TLS keys are optional, and if present, will override the
default XCM TLS credential file paths.

Old-style server configuration files are still valid.

The test suite is adapted to verify the new functionality.

Tweaks are made to the test suite to run more test cases with both
file-based and command line-based configuration.

Resolves #25.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>