Update XCM wrapper
This patch brings in a new version of the XCM wrapper from the XCM
repo.

The new version includes bugfixes and support for binary and
double-type XCM socket attributes.

The XCM version API wrappers are removed, to allow it to function
properly with XCM version prior to v1.5.0.

The xcm_want()-related parts are also removed, since they are not used
by Pathfinder.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>