Add server-side certificate revocation list checking
Introduced support for CRL checking in pafd.

This commit adds a new key 'crl' in the domain's socket list
configuration. If set, CRL checking will be enabled.

The server tests are modified to use the 'gencert' script brought in
from the XCM repo to generate the certificate, keys, and CRLs. This
change makes test credential generation much more readable, compared
to the prior OpenSSL-based approach.

The README is updated to describe the dependency to XCM v1.9.0 if (and
only if) CRL checking functionality is configured in the server.

This commit also fixes an issue with TLS attribute test coverage. The
tests were set up so that the server could fall back on a working
XCM_TLS_CERT setting. This hid an issue in the TLS attribute test
cases where TLS attribute tests occasionally were run with a setup
where no server configuration file was employed (but only command-line
arguments), and thus no TLS attribute could possibly be configured.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>