Fix unpublish non-orphan republished service bug
Fix bug which caused a server crash in a scenario where a service was
published by one client, republished by another client with the
service still be an non-orphan (i.e., the original client had not
left), and then unpublished.

A test case to reproduce the issue is added.

This code was found during code review.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>