Improve pafc track command
The pafc track command syntax is remodeled to fit a knob to turn
logging server queries on. Default is now off.

If a low-TTL service is published, the server will send frequent track
queries, which makes the pafc CLI hard to use, in case each reply is
logged to the console.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>