Tune idle time parameters
Increase minimum max idle time from 2 to 4 seconds, and send query
already when 50% of the idle time has expired, not 75%. This will
translate to a max allowed track latency going from 500 ms to 2
seconds.

The goal of this change is to increase robustness when clients,
servers or the network are not very responsive.

This change will increase the load generated by high-TTL clients, and
slightly lower the load generated by clients which has published at
least one low-TTL service.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>