Allow limiting protocol versions used
Support configuring an upper and/or lower bound of the protocol
version allowed by the server for clients connecting to a particular
domain. The limits are specified in the server's configuration file.

The pafd manual page and the example server configuration file is
updated, and test cases verifying the server behavior are added.

Resolves #49.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>