Republish orphan service before unpublishing
In case an orphan service is unpublished (currently owned by the same
client id, or a different), the unpublishing client should take
ownership of the service, and make it a non-orphan.

This allows for any consumers to (via a subscription) notice that this
is an explicit unpublish operation, and act accordingly.

The both v2 and v3 protocol specifications are updated to reflect this
change. These changes are deemed to be backward compatible.

Resolves #52.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>